move files and keep original
