--- optolink_splitter/optolinkvs2_switch_bkp.py
@@ -1,363 +1,0 @@
-"""
-   Copyright 2024 philippoo66
-   
-   Licensed under the GNU GENERAL PUBLIC LICENSE, Version 3 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       https://www.gnu.org/licenses/gpl-3.0.html
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-"""
-
-import serial
-import time
-import threading
-
-
-from optolink_splitter.config_model import SplitterConfig
-from optolink_splitter.optolinkvs2 import init_vs2, receive_vs2telegr
-from optolink_splitter.utils.common_utils import csv_to_tuple_list, bbbstr
-from optolink_splitter.utils.mqtt_util import (
-    connect_mqtt,
-    publish_read,
-    get_mqtt_request,
-    publish_response,
-    exit_mqtt,
-)
-from optolink_splitter.utils.requests_util import (
-    response_to_request,
-    perform_bytebit_filter,
-)
-from optolink_splitter.utils.tcpip_util import (
-    tcpip4ever,
-    get_tcp_request,
-    send_tcpip,
-    exit_tcpip,
-)
-from optolink_splitter.utils.viessdata_util import buffer_csv_line
-from optolink_splitter.utils.viconn_util import (
-    detect_vs2,
-    listen_to_Vitoconnect,
-    get_vicon_request,
-)
-
-# global_exit_flag = False
-vitolog = None  # Vitoconnect logging
-poll_pointer = 0  # polling list
-timer_pollinterval = threading.Timer(1, None)  # Initialise global timer pollinterval
-
-
-def olbreath(retcode: int) -> None:
-    if retcode <= 0x03:
-        # success, err msg
-        time.sleep(0.1)
-    elif retcode in [0xFF, 0xAA]:
-        # timeout, err_handle
-        pass
-    else:
-        # allow calming down
-        time.sleep(0.5)
-
-
-def log_vito(data, format_data_hex_format: str, pre):
-    global vitolog
-    if vitolog is not None:
-        sd = bbbstr(data, format_data_hex_format)
-        vitolog.write(f"{pre}\t{int(time.time()*1000)}\t{sd}\n")
-
-
-def do_poll_item(
-    config: SplitterConfig,
-    poll_items: list[tuple],
-    w1sensors: list[tuple],
-    poll_data,
-    ser: serial.Serial,
-    mqtt_address=None,
-) -> int:  # retcode
-    global poll_pointer
-    val = "?"
-
-    item = poll_items[poll_pointer]  # (Name, DpAddr, Len, Scale/Type, Signed)
-    retcode, data, val, _ = response_to_request(config, w1sensors, item, ser)
-
-    if retcode == 0x01:
-        # save val in buffer for csv
-        poll_data[poll_pointer] = val
-
-        # post to MQTT broker
-        if mqtt_address is not None:
-            publish_read(config, item[0], item[1], val)
-
-        # probably more bytebit values of the same datapoint?!
-        if len(item) > 3:
-            if str(item[3]).lower().startswith("b:"):
-                # bytebit filter +++++++++
-                while (poll_pointer + 1) < len(poll_items):
-                    next_idx = poll_pointer + 1
-                    next_item = poll_items[next_idx]
-                    # if next address same AND next len same AND next type starts with 'b:'
-                    if (
-                        (next_item[1] == item[1])
-                        and (next_item[2] == item[2])
-                        and (str(next_item[3]).lower()).startswith("b:")
-                    ):
-                        next_val = perform_bytebit_filter(config, data, next_item)
-
-                        # save val in buffer for csv
-                        poll_data[next_idx] = next_val
-
-                        if mqtt_address is not None:
-                            # post to MQTT broker
-                            publish_read(config, next_item[0], next_item[1], next_val)
-
-                        poll_pointer = next_idx
-                    else:
-                        break
-    else:
-        print(
-            f"Error do_poll_item {poll_pointer}, Addr {item[1]:04X}, RetCode {retcode}, Data {val}"
-        )
-    return retcode
-
-
-# poll timer
-def on_polltimer(poll_items: list[tuple], poll_interval: int) -> None:
-    global poll_pointer
-    poll_pointer = (poll_pointer + 1) % len(poll_items)
-    startPollTimer(poll_items, poll_interval)
-
-
-def startPollTimer(poll_items: list[tuple], poll_interval: int) -> None:
-    global timer_pollinterval
-    if timer_pollinterval is not None:
-        timer_pollinterval.cancel()
-    timer_pollinterval = threading.Timer(
-        poll_interval, on_polltimer, args=(poll_items, poll_interval)
-    )
-    timer_pollinterval.start()
-
-
-def optolink_vs2_switch(config: SplitterConfig) -> None:
-    global poll_pointer
-    global vitolog
-
-    poll_items = csv_to_tuple_list(config.poll_items_config_path)
-    w1sensors = None
-    if config.w1sensors_config_path is not None:
-        w1sensors = csv_to_tuple_list(config.w1sensors_config_path)
-
-    try:
-        mod_mqtt_util = None
-        poll_data = [None] * len(poll_items)
-
-        # serielle Verbidungen mit Vitoconnect und dem Optolink Kopf aufbauen ++++++++++++++
-        serViCon = None  # Vitoconnect (Master)
-        serViDev = None  # Viessmann Device (Slave)
-
-        if config.vitoconnect_port is not None:
-            serViCon = serial.Serial(
-                config.vitoconnect_port,
-                baudrate=4800,
-                parity=serial.PARITY_EVEN,
-                stopbits=serial.STOPBITS_TWO,
-                bytesize=serial.EIGHTBITS,
-                exclusive=True,
-                timeout=0,
-            )
-
-        if config.optolink_port is not None:
-            serViDev = serial.Serial(
-                config.optolink_port,
-                baudrate=4800,
-                parity=serial.PARITY_EVEN,
-                stopbits=serial.STOPBITS_TWO,
-                bytesize=serial.EIGHTBITS,
-                exclusive=True,
-                timeout=0,
-            )
-        else:
-            raise Exception("Error: Optolink device is mandatory!")
-
-        # Empfangstask der sekundären Master starten (TcpIp, MQTT)
-
-        # MQTT --------
-        if config.mqtt_address is not None:
-            connect_mqtt(config)
-
-        # TCP/IP connection --------
-        if config.tcpip_port is not None:
-            tcp_thread = threading.Thread(
-                target=tcpip4ever, args=(config, poll_items, config.tcpip_port, False)
-            )
-            tcp_thread.daemon = (
-                True  # Setze den Thread als Hintergrundthread - wichtig für Ctrl-C
-            )
-            tcp_thread.start()
-
-        # run VS2 connection ------------------
-        if serViCon is not None:
-            # Vitoconncet logging
-            if config.logging_vitoconnect_log_path is not None:
-                vitolog = open(config.logging_vitoconnect_log_path, "a")
-            # detect VS2 Protokol
-            print("awaiting VS2...")
-            vs2timeout = config.vitoconnect_vs2timeout
-            if not detect_vs2(config, serViCon, serViDev, vs2timeout, vitolog):
-                raise Exception("VS2 protocol not detected within timeout", vs2timeout)
-            print("VS detected")
-            vicon_thread = threading.Thread(
-                target=listen_to_Vitoconnect, args=(config, serViCon, vitolog)
-            )
-            vicon_thread.daemon = (
-                True  # Setze den Thread als Hintergrundthread - wichtig für Ctrl-C
-            )
-            vicon_thread.start()
-        else:
-            # VS2 Protokoll am Slave initialisieren
-            if not init_vs2(serViDev):
-                print("init_vs2 failed")
-                raise Exception("init_vs2 failed")  # schlecht für KW Protokoll
-
-        # Polling Mechanismus --------
-        len_polllist = len(poll_items)
-        if (config.poll_interval > 0) and (len_polllist > 0):
-            startPollTimer(poll_items, config.poll_interval)
-
-        # Main Loop starten und Sachen abarbeiten
-        request_pointer = 0
-        while True:
-            tookbreath = False
-            if serViCon is not None:
-                # first Vitoconnect request -------------------
-                vidata = get_vicon_request()
-                if vidata:
-                    serViDev.write(vidata)
-                    log_vito(vidata, config.format_data_hex_format, "M")
-                    # recive response an pass bytes directly back to VitoConnect,
-                    # returns when response is complete (or error or timeout)
-                    retcode, _, redata = receive_vs2telegr(
-                        config.format_data_hex_format,
-                        config.logging_show_opto_rx,
-                        True,
-                        True,
-                        serViDev,
-                        serViCon,
-                    )
-                    log_vito(redata, config.format_data_hex_format, "S")
-                    olbreath(retcode)
-                    tookbreath = True
-
-            # secondary requests ------------------
-            # TODO überlegen/testen, ob Vitoconnect request nicht auch in der Reihe reicht
-
-            # polling list --------
-            if request_pointer == 0:
-                if config.poll_interval < 0:
-                    request_pointer += 1
-                elif poll_pointer < len_polllist:
-                    retcode = do_poll_item(
-                        config,
-                        poll_items,
-                        w1sensors,
-                        poll_data,
-                        serViDev,
-                        config.mqtt_address,
-                    )
-
-                    poll_pointer += 1
-
-                    if poll_pointer == len_polllist:
-                        if config.viessdata_csv_path is not None:
-                            buffer_csv_line(
-                                poll_data,
-                                config.viessdata_csv_path,
-                                config.viessdata_csv_buffer_to_write,
-                                config.viessdata_csv_delimiter,
-                                poll_items,
-                            )
-                        poll_pointer += 1
-                        if config.poll_interval == 0:
-                            poll_pointer = 0
-                    olbreath(retcode)
-                    tookbreath = True
-                else:
-                    request_pointer += 1
-
-            # MQTT request --------
-            if request_pointer == 1:
-                if config.mqtt_address is None:
-                    request_pointer += 1
-                else:
-                    msg = get_mqtt_request()
-                    if msg:
-                        try:
-                            retcode, _, _, resp = response_to_request(
-                                config, w1sensors, msg, serViDev
-                            )
-                            publish_response(config, resp)
-                            olbreath(retcode)
-                            tookbreath = True
-                        except Exception as e:
-                            print("Error handling MQTT request:", e)
-                    else:
-                        request_pointer += 1
-
-            # TCP/IP request --------
-            if request_pointer == 2:
-                if config.tcpip_port is None:
-                    request_pointer += 1
-                else:
-                    msg = get_tcp_request()
-                    if msg:
-                        try:
-                            retcode, _, _, resp = response_to_request(
-                                config, w1sensors, msg, serViDev
-                            )
-                            send_tcpip(resp)
-                            olbreath(retcode)
-                            tookbreath = True
-                        except Exception as e:
-                            print("Error handling TCP request:", e)
-                    else:
-                        request_pointer += 1
-
-            # request_pointer control --------
-            request_pointer += 1
-            if request_pointer > 2:
-                request_pointer = 0
-
-            # let cpu take a breath
-            if not tookbreath:
-                time.sleep(0.005)
-
-    except KeyboardInterrupt:
-        print("Abbruch durch Benutzer.")
-    except Exception as e:
-        print(e)
-    finally:
-        # sauber beenden: Tasks stoppen, VS1 Protokoll aktivieren(?), alle Verbindungen trennen
-        # Schließen der seriellen Schnittstellen, Ausgabedatei, PollTimer,
-        print("exit close")
-        if serViCon is not None:
-            print("closing serViCon")
-            serViCon.close()
-        if serViDev is not None:
-            if serViDev.is_open:
-                print("reset protocol")
-                serViDev.write([0x04])  # TODO yes or no?
-                print("closing serViDev")
-                serViDev.close()
-        print("cancel poll timer ")
-        timer_pollinterval.cancel()
-        exit_tcpip()
-        # tcp_thread.join()  #TODO ??
-        if config.mqtt_address is not None:
-            exit_mqtt()
-        if vitolog is not None:
-            print("closing vitolog")
-            vitolog.close()
